+++ guillotina/interfaces/__init__.py
@@ -20,7 +20,6 @@ from .content import ILocation  # noqa
 from .content import IRegistry  # noqa
 from .content import IResource  # noqa
 from .content import IResourceFactory  # noqa
-from .content import IContainer  # noqa
 from .content import IStaticDirectory  # noqa
 from .content import IStaticFile  # noqa
 from .content import ITraversable  # noqa
@@ -29,6 +28,7 @@ from .events import IBeforeObjectRemovedEvent  # noqa
 from .events import IFileFinishUploaded  # noqa
 from .events import INewUserAdded  # noqa
 from .events import IObjectAddedEvent  # noqa
+from .events import IObjectLoadedEvent  # noqa
 from .events import IObjectModifiedEvent  # noqa
 from .events import IObjectMovedEvent  # noqa
 from .events import IObjectPermissionsModifiedEvent  # noqa