+++ guillotina/interfaces/events.py
@@ -32,6 +32,10 @@ class IBeforeObjectRemovedEvent(IObjectMovedEvent):
     """An object has been removed from a container."""
 
 
+class IObjectLoadedEvent(interfaces.IObjectEvent):
+    """An objects has been loaded from the database"""
+
+
 class IObjectVisitedEvent(interfaces.IObjectEvent):
     """An object has been vIContainerd."""